feat: function without chat

+++ lib/screen/part1/signup_screen2.dart
@@ -1,10 +1,13 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 import 'package:shared_preferences/shared_preferences.dart';
+import 'package:smile_front/screen/part2/home_screen.dart';
 import '/firebase_options.dart';
+import 'package:intl/intl.dart';
 
 import './depression_diagnosis_selection_screen.dart';
 
@@ -36,15 +39,121 @@ class _SignupScreenState extends State<SignupScreen> {
   late String email;
   late String password;
   late String name;
+  late int flag = 0;
+  late int userScore;
+  late List<String> medications;
+  late bool _complete1 = false;
+  late bool _complete2 = false;
+  late bool _complete3 = false;
+  late bool _complete4 = false;
+  late bool _complete5 = false;
+  late bool _complete6 = false;
+
+  late String _dt1 = "";
+  late String _dt2 = "";
+  late String _dt3 = "";
+  late String _dt4 = "";
+  late String _dt5 = "";
+  late String _dt6 = "";
+  late int _mission1 = 0;
+  late int _mission2 = 0;
+  late int _mission3 = 0;
+  late int _mission4 = 0;
+  late int _mission5 = 0;
+  late int _mission6 = 0;
   bool isInput = true; //false - result
   bool isSignIn = true; //false - SingUp
+
+  getUser() async {
+    final DateTime now = DateTime.now();
+    final String nowStr = DateFormat('yy-MM-dd').format(now);
+    DocumentSnapshot _userDoc = await FirebaseFirestore.instance.collection('users').doc(email).get();
+    if (_userDoc.exists) {
+      Map<String, dynamic> userData = _userDoc.data() as Map<String, dynamic>;
+      email = userData['userInfo'];
+      name = userData['userName'];
+      userScore = userData['userScore'];
+      medications = List<String>.from(userData['medications']);
+      _complete1 = userData['complete1'] ?? false;
+      _complete2 = userData['complete2'] ?? false;
+      _complete3 = userData['complete3'] ?? false;
+      _complete4 = userData['complete4'] ?? false;
+      _complete5 = userData['complete5'] ?? false;
+      _complete6 = userData['complete6'] ?? false;
+      _dt1 = userData['dt1'] ?? "";
+      _dt2 = userData['dt2'] ?? "";
+      _dt3 = userData['dt3'] ?? "";
+      _dt4 = userData['dt4'] ?? "";
+      _dt5 = userData['dt5'] ?? "";
+      _dt6 = userData['dt6'] ?? "";
+      _mission1 = userData['mission1'] ?? 0;
+      _mission2 = userData['mission2'] ?? 0;
+      _mission3 = userData['mission3'] ?? 0;
+      _mission4 = userData['mission4'] ?? 0;
+      _mission5 = userData['mission5'] ?? 0;
+      _mission6 = userData['mission6'] ?? 0;
+
+      SharedPreferences _prefs = await SharedPreferences.getInstance();
+      _prefs.setString('userInfo', email);
+      _prefs.setString('userName', name);
+      _prefs.setInt('userScore', userScore);
+      _prefs.setStringList('medications', medications);
+      if (nowStr == _dt1){
+        _prefs.setBool('complete1', _complete1);
+      } else {
+        _prefs.setBool('complete1', false);
+      }
+      if (nowStr == _dt2){
+        _prefs.setBool('complete2', _complete2);
+      } else {
+        _prefs.setBool('complete2', false);
+      }
+      if (nowStr == _dt3){
+        _prefs.setBool('complete3', _complete3);
+      } else {
+        _prefs.setBool('complete3', false);
+      }
+      if (nowStr == _dt4){
+        _prefs.setBool('complete4', _complete4);
+      } else {
+        _prefs.setBool('complete4', false);
+      }
+      if (nowStr == _dt5){
+        _prefs.setBool('complete5', _complete5);
+      } else {
+        _prefs.setBool('complete5', false);
+      }
+      if (nowStr == _dt6){
+        _prefs.setBool('complete6', _complete6);
+      } else {
+        _prefs.setBool('complete6', false);
+      }
+      _prefs.setString('dt1', _dt1);
+      _prefs.setString('dt2', _dt2);
+      _prefs.setString('dt3', _dt3);
+      _prefs.setString('dt4', _dt4);
+      _prefs.setString('dt5', _dt5);
+      _prefs.setString('dt6', _dt6);
+      _prefs.setInt('mission1', _mission1);
+      _prefs.setInt('mission2', _mission2);
+      _prefs.setInt('mission3', _mission3);
+      _prefs.setInt('mission4', _mission4);
+      _prefs.setInt('mission5', _mission5);
+      _prefs.setInt('mission6', _mission6);
+
+      flag = 1;
+    } else {
+      flag = 0;
+    }
+  }
 
   signIn() async {
     try {
       await FirebaseAuth.instance
           .signInWithEmailAndPassword(email: email, password: password)
           .then((value) {
-        print(value);
+        print(value); // value에 사용자 이메일 들어가 있음.
+
         if (value.user!.emailVerified) {
           //이메일 인증 여부
           setState(() {
@@ -153,9 +262,11 @@ class _SignupScreenState extends State<SignupScreen> {
             ElevatedButton(onPressed: (){
               if (_formKey.currentState?.validate() ?? false) {
                 _formKey.currentState?.save();
-                print('email: $email, password : $password');
+                // print('email: $email, password : $password');
                 if (isSignIn) {
                   signIn();
+                  getUser();
+                  print(flag);
                 } else {
                   signUp();
                 }
@@ -203,16 +314,23 @@ class _SignupScreenState extends State<SignupScreen> {
       ),
       ElevatedButton(
           onPressed: () async {
-            if (isSignIn) {
+            if (isSignIn && flag == 0) {
               SharedPreferences prefs = await SharedPreferences.getInstance();
               prefs.clear();
-              prefs.setString('email', resultEmail);
-              print(prefs.get('email'));
+              prefs.setString('userInfo', resultEmail);
+              prefs.setString('userName', resultEmail.split('@')[0]);
+              print(prefs.get('userInfo'));
               Navigator.push(
                 context,
                 MaterialPageRoute(builder: (context) => DepressionDiagnosisSelectionScreen()),
               );
-            } else {
+            } else if (isSignIn && flag == 1) {
+              Navigator.push(
+                context,
+                MaterialPageRoute(builder: (context) => HomeScreen()),
+              );
+            }
+            else {
               setState(() {
                 isInput = true;
                 isSignIn = true;
